Add java servlets to process requests for submitting and deleting markers
