+++ portfolio/src/main/webapp/script.js
@@ -115,150 +115,3 @@ function deleteComments() {
     method: 'POST',
   }).then(res => console.log("Delete all comments"));
 }
-
-function initMap() {
-  // Create styled map.
-  var styledMapType = new google.maps.StyledMapType(
-    [
-      {elementType: 'geometry', stylers: [{color: '#ebe3cd'}]},
-      {elementType: 'labels.text.fill', stylers: [{color: '#523735'}]},
-      {elementType: 'labels.text.stroke', stylers: [{color: '#f5f1e6'}]},
-      {
-        featureType: 'administrative',
-        elementType: 'geometry.stroke',
-        stylers: [{color: '#c9b2a6'}]
-      },
-      {
-        featureType: 'administrative.land_parcel',
-        elementType: 'geometry.stroke',
-        stylers: [{color: '#dcd2be'}]
-      },
-      {
-        featureType: 'administrative.land_parcel',
-        elementType: 'labels.text.fill',
-        stylers: [{color: '#ae9e90'}]
-      },
-      {
-        featureType: 'landscape.natural',
-        elementType: 'geometry',
-        stylers: [{color: '#dfd2ae'}]
-      },
-      {
-        featureType: 'poi',
-        elementType: 'geometry',
-        stylers: [{color: '#dfd2ae'}]
-      },
-      {
-        featureType: 'poi',
-        elementType: 'labels.text.fill',
-        stylers: [{color: '#93817c'}]
-      },
-      {
-        featureType: 'poi.park',
-        elementType: 'geometry.fill',
-        stylers: [{color: '#a5b076'}]
-      },
-      {
-        featureType: 'poi.park',
-        elementType: 'labels.text.fill',
-        stylers: [{color: '#447530'}]
-      },
-      {
-        featureType: 'road',
-        elementType: 'geometry',
-        stylers: [{color: '#f5f1e6'}]
-      },
-      {
-        featureType: 'road.arterial',
-        elementType: 'geometry',
-        stylers: [{color: '#fdfcf8'}]
-      },
-      {
-        featureType: 'road.highway',
-        elementType: 'geometry',
-        stylers: [{color: '#f8c967'}]
-      },
-      {
-        featureType: 'road.highway',
-        elementType: 'geometry.stroke',
-        stylers: [{color: '#e9bc62'}]
-      },
-      {
-        featureType: 'road.highway.controlled_access',
-        elementType: 'geometry',
-        stylers: [{color: '#e98d58'}]
-      },
-      {
-        featureType: 'road.highway.controlled_access',
-        elementType: 'geometry.stroke',
-        stylers: [{color: '#db8555'}]
-      },
-      {
-        featureType: 'road.local',
-        elementType: 'labels.text.fill',
-        stylers: [{color: '#806b63'}]
-      },
-      {
-        featureType: 'transit.line',
-        elementType: 'geometry',
-        stylers: [{color: '#dfd2ae'}]
-      },
-      {
-        featureType: 'transit.line',
-        elementType: 'labels.text.fill',
-        stylers: [{color: '#8f7d77'}]
-      },
-      {
-        featureType: 'transit.line',
-        elementType: 'labels.text.stroke',
-        stylers: [{color: '#ebe3cd'}]
-      },
-      {
-        featureType: 'transit.station',
-        elementType: 'geometry',
-        stylers: [{color: '#dfd2ae'}]
-      },
-      {
-        featureType: 'water',
-        elementType: 'geometry',
-        stylers: [{color: '#17263c'}]
-      },
-      {
-        featureType: 'water',
-        elementType: 'labels.text.fill',
-        stylers: [{color: '#515c6d'}]
-      },
-      {
-        featureType: 'water',
-        elementType: 'labels.text.stroke',
-        stylers: [{color: '#17263c'}]
-      }
-    ],
-    {name: 'Styled Map'}
-  );
-
-  // Create markers.
-  let japanTemple = {lat: 34.462117, lng: 135.830272};
-  let newZealandLake = {lat: -43.979931, lng: 170.194799};
-  let ugandaView = {lat: -0.099273, lng: 32.652921}
-
-  // Create a map object, and include the MapTypeId to add
-  // to the map type control.
-  let map = new google.maps.Map(document.getElementById('map'), {
-    center: new google.maps.LatLng(0.211772, 102.290621),
-    zoom: 2.3,
-    mapTypeControlOptions: {
-      mapTypeIds: ['roadmap', 'satellite', 'hybrid', 'terrain',
-              'styled_map']
-    }
-  });
-
-  //Associate the styled map with the MapTypeId and set it to display.
-  map.mapTypes.set('styled_map', styledMapType);
-  map.setMapTypeId('styled_map');
-
-  // The marker, positioned at Uluru
-  let markerJapanTemple = new google.maps.Marker({position: japanTemple, map: map});
-  let markerNewZealand = new google.maps.Marker({position: newZealandLake, map: map});
-  let markerUganda = new google.maps.Marker({position: ugandaView, map: map});
-}
